chore(purchase): implement delete and get of purchase, item purchase and supplier use case with tests Added: delete-supplier, delete-purchase, delete-item-purchase, get-item-purchase, get-purchase, get-supplier use case and their respective tests. Note: Controllers still need to be implemented.

+++ src/purchase/application/usecases/delete-purchase.ts
@@ -1,4 +1,5 @@
 import { Inject, Injectable } from '@nestjs/common';
+import { ItemPurchaseRepository } from 'src/purchase/domain/repository/item-purchase.repository';
 import { PurchaseRepository } from 'src/purchase/domain/repository/purchase.repository';
 
 @Injectable()
@@ -6,10 +7,22 @@ export class DeletePurchase {
     constructor(
         @Inject('PurchaseRepository')
         private readonly purchaseRepository: PurchaseRepository,
+        @Inject('ItemPurchaseRepository')
+        private readonly itemPurchaseRepository: ItemPurchaseRepository,
     ) {}
 
     async execute(input: Input): Promise<Output> {
         const { id } = input;
+
+        const purchase = await this.purchaseRepository.get(id);
+        if (!purchase) {
+            throw new Error('Purchase not found');
+        }
+
+        const items = await this.itemPurchaseRepository.getAllByPurchaseId(id);
+        if (items.length > 0) {
+            throw new Error('Cannot delete purchase with items');
+        }
 
         await this.purchaseRepository.delete(id);
     }
